backend/requirements.txt
fastapi
uvicorn
sqlalchemy
cryptography
passlib[bcrypt]
python-dotenv
email-validator
